// backend/src/common/enums/permission.enum.ts
export enum PermissionEnum {
  //admin permission
  SET_PERMISSION = 1,
  VIEW_ALL_EMPLOYEE = 2,
  CREATE_EMPLOYEE = 3,
  DELETE_EMPLOYEE = 4,
  VIEW_ALL_DEPARTMENT = 5,
  REMOVE_EMPLOYEE_ALL_DEPARTMENT = 6,
  INVITE_EMPLOYEE_ALL_DEPARTMENT = 7,
  CREATE_DEPARTMENT = 8,
  UPDATE_DEPARTMENT_ALL_DEPARTMENT = 9,
  DELETE_DEPARTMENT_ALL_DEPARTMENT = 10,
  VIEW_ATTENDANCE_ALL_DEPARTMENT = 11, //2
  VIEW_LEAVE_REQUEST_ALL_DEPARTMENT = 12, //2
  APPROVE_LEAVE_REQUEST_ALL_DEPARTMENT = 13, //2
  REJECT_LEAVE_REQUEST_ALL_DEPARTMENT = 14, //2
}